+++ Makefile
@@ -21,6 +21,7 @@ DIR2		:= $(SRCDIR)/input
 DIR3		:= $(SRCDIR)/wrapper
 DIR4		:= $(SRCDIR)/parse
 DIR5		:= $(SRCDIR)/utils
+DIR6		:= $(SRCDIR)/draw
 ALLDIRS		:= $(shell find srcs -mindepth 1 -type d)
 
 MAIN		:= srcs/main.c\
@@ -49,7 +50,11 @@ SRC4	=\
 SRC5	=\
 	srcs/utils/utils.c\
 
-SRCS		:= $(MAIN)$(SRC1)$(SRC2)$(SRC3)$(SRC4)$(SRC5)
+SRC6	=\
+	srcs/draw/draw.c\
+	srcs/draw/fill_screen.c\
+
+SRCS		:= $(MAIN)$(SRC1)$(SRC2)$(SRC3)$(SRC4)$(SRC5)$(SRC6)
 OBJS		:= $(patsubst $(SRCDIR)%,$(OBJDIR)%,$(SRCS:.c=.o))
 OBJSUBDIRS	:= $(patsubst $(SRCDIR)%,$(OBJDIR)%,$(ALLDIRS))
 VPATH 		:= $(shell find $(SRCDIR) -type d | tr '\n' ':')
@@ -112,11 +117,13 @@ add		:
 	bash header.sh "$(DIR3)" $(INCLUDE)/wrapper.h
 	bash header.sh "$(DIR4)" $(INCLUDE)/parse.h
 	bash header.sh "$(DIR5)" $(INCLUDE)/utils.h
+	bash header.sh "$(DIR6)" $(INCLUDE)/draw.h
 	bash make.sh $(DIR1) SRC1
 	bash make.sh $(DIR2) SRC2
 	bash make.sh $(DIR3) SRC3
 	bash make.sh $(DIR4) SRC4
 	bash make.sh $(DIR5) SRC5
+	bash make.sh $(DIR6) SRC6
 
 re		: fclean all
