+++ libft/srcs/utils/ft_max_dbl.c
@@ -1,4 +1,3 @@
-
 double	ft_max_dbl(double a, double b)
 {
 	if (a > b)
