+++ libft/srcs/utils/ft_min_dbl.c
@@ -1,4 +1,3 @@
-
 double	ft_min_dbl(double a, double b)
 {
 	if (a > b)
